Demo GUI is ready
