--- URL-analyzer/website/__init__.py
@@ -1,12 +1,0 @@
-from flask import Flask
-
-
-def create_app():
-    app = Flask(__name__)
-    app.config['SECRET_KEY'] = 'asdsdgtrbfgbrdfsdf'
-
-    from .views import views
-
-    app.register_blueprint(views, url_prefix='/')
-
-    return app
